move aws init logic into separate lib (#41093)
