+++ ydb/core/driver_lib/run/ya.make
@@ -141,6 +141,7 @@ PEERDIR(
     ydb/library/actors/protos
     ydb/library/actors/retro_tracing
     ydb/library/actors/util
+    ydb/library/aws_init
     ydb/library/folder_service
     ydb/library/folder_service/proto
     ydb/library/global_plugins
