+++ ydb/core/kqp/ut/olap/ya.make
@@ -34,6 +34,7 @@ PEERDIR(
     ydb/core/kqp/ut/olap/helpers
     ydb/core/kqp/ut/olap/combinatory
     ydb/core/tx/datashard/ut_common
+    ydb/library/aws_init
     ydb/public/sdk/cpp/src/client/operation
 )
