+++ ydb/core/tx/schemeshard/ut_export_reboots_s3/ya.make
@@ -24,6 +24,7 @@ PEERDIR(
     ydb/core/tx/schemeshard/ut_helpers
     ydb/core/util
     ydb/core/wrappers/ut_helpers
+    ydb/library/aws_init
     yql/essentials/public/udf/service/exception_policy
 )
