+++ ydb/core/tx/schemeshard/ut_index_build/ut_vector_index_build.cpp
@@ -10,7 +10,7 @@
 #include <ydb/core/metering/metering.h>
 
 #include <ydb/core/wrappers/ut_helpers/s3_mock.h>
-#include <ydb/core/util/aws.h>
+#include <ydb/library/aws_init/aws.h>
 #include <ydb/public/api/protos/ydb_import.pb.h>
 
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/table/table.h>
