+++ ydb/core/tx/schemeshard/ut_index_build/ya.make
@@ -14,6 +14,7 @@ PEERDIR(
     ydb/core/tx
     ydb/core/tx/schemeshard/ut_helpers
     ydb/core/wrappers/ut_helpers
+    ydb/library/aws_init
     ydb/public/sdk/cpp/src/client/table
 )
