+++ ydb/core/util/ya.make
@@ -84,23 +84,6 @@ PEERDIR(
     ydb/library/yverify_stream
 )
 
-IF (OS_WINDOWS)
-    CFLAGS(
-        -DKIKIMR_DISABLE_S3_OPS
-    )
-    SRCS(
-        aws_windows_stub.cpp
-    )
-ELSE()
-    PEERDIR(
-        contrib/libs/aws-sdk-cpp/aws-cpp-sdk-core
-        contrib/libs/curl
-    )
-    SRCS(
-        aws.cpp
-    )
-ENDIF()
-
 END()
 
 RECURSE_FOR_TESTS(
