+++ ydb/core/wrappers/ut/ya.make
@@ -13,6 +13,7 @@ IF (NOT OS_WINDOWS)
         ydb/core/testlib/basics/default
         ydb/core/util
         ydb/core/wrappers/ut_helpers
+        ydb/library/aws_init
         yql/essentials/minikql/comp_nodes/llvm16
     )
     SRCS(
