+++ ydb/library/aws_init/aws_windows_stub.cpp
